+++ G1-ee-groep1-palamedes.SH-MVL.API/Models/Art.cs
@@ -5,7 +5,7 @@ namespace G1_ee_groep1_palamedes.SH_MVL.API.Models
     public class Art : EntityBase
     {
         public string Name { get; set; }
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
         public Category Category { get; set; }
         public DateTime Created { get; set; }
         public string Description { get; set; }
@@ -13,7 +13,7 @@ namespace G1_ee_groep1_palamedes.SH_MVL.API.Models
         public decimal Price { get; set; }
         public string ImageName { get; set; }
 
-        public int ArtistId { get; set; }
+        public long ArtistId { get; set; }
         public Artist Artist { get; set; }
 
     }
